Fix memory leaks
- Fix a variety of memory leaks by freeing menu items, temporary
  variables, and query results

Note: Main menu item userptrs is now a duplicate string, making it
freeable by the free_menu_items() function. This also means we can call
PQclear() immediately after populating the menu.

Bug: This commit introduces a bug in creating a person summary, in that
postgres apparently doesn't receive a value for the UUID. This bug seems
to go away if I add in some print statements in random places, so it'll
be a fun one to find.