Fix bug preventing person summaries from opening
- Duplicate userptr string in main_menu delegate before passing it back
  to the caller
- Free *uid string in person_summary after use
- Delete debug print statement in persom_summary delegate

As the userptr gets freed by free_menu_items(), the uid string was set
to an empty string before being passed back to the caller. I'm not sure
why it became a readable empty string, as freeing the char* should have
made any subsequent read attempts result in a segfault.